requirements.txt
pybaseball==2.2.7
xgboost==2.1.3
shap==0.46.0
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.2.4
mlb-statsapi==1.9.0
requests==2.32.3
dash==2.18.2
plotly==5.24.1
openpyxl==3.1.5
python-dotenv==1.0.1
schedule==1.2.2
joblib==1.4.2
imbalanced-learn==0.13.0
